requirements.txt
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==1.0.1
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
